[fix] imported from local

+++ connect.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Log levels
+# Log levels, enums and constants
 readonly LOG_LEVEL_ERROR="ERROR"
 readonly LOG_LEVEL_INFO="INFO"
 readonly LOG_LEVEL_DEBUG="DEBUG"
@@ -11,11 +11,22 @@ readonly APP_AVAILALBE_TO_OTHER_TENANTS="false"
 readonly SECRETS_STORE="/tmp/spot_azure_secrets_store.json"
 readonly ACCOUNTS_STORE="/tmp/spot_azure_accounts_store.json"
 readonly TEMP_APP_DETAILS_LOCATION="/tmp/spot_app_azure_details.json"
-readonly ON_FAILURE_CLEANUP_FOLDER="/tmp/spot_azure_on_failure.json"
 readonly RESOURCE_FILE="/tmp/spot_azure_created_resources.json"
 readonly RESOURCE_FILE_TEMP="/tmp/spot_azure_created_resources_temp.json"
 readonly NO_SIGNAL_SUCCESS="NO_SIGNAL_SUCCESS"
 readonly EXISTING_ROLE_ASSIGNMENT_REFERENCE_ON_DELETION="There are existing role assignments referencing"
+readonly SIGNAL_APP_REGISTRATION_CREATION_ENUM="SIGNAL_APP_REGISTRATION_CREATION"
+readonly SIGNAL_IAM_ROLE_CREATION_ENUM="SIGNAL_IAM_ROLE_CREATION"
+readonly SIGNAL_ROLE_ASSIGNMENT_SUCCESS_ENUM="SIGNAL_ROLE_ASSIGNMENT_SUCCESS"
+readonly SIGNAL_SPOT_AZURE_SUBSCRIPTION_CONNECTED_SUCCESS_ENUM="SIGNAL_SPOT_AZURE_SUBSCRIPTION_CONNECTED_SUCCESS"
+readonly SIGNAL_SERVICE_PRINCIPAL_CREATION_ENUM="SIGNAL_SERVICE_PRINCIPAL_CREATION"
+readonly SIGNAL_APPLICATION_SECRET_RESET_ENUM="SIGNAL_APPLICATION_SECRET_RESET"
+export SIGNAL_APP_REGISTRATION_CREATION=0
+export SIGNAL_APPLICATION_SECRET_RESET=0
+export SIGNAL_IAM_ROLE_CREATION=0
+export SIGNAL_ROLE_ASSIGNMENT_SUCCESS=0
+export SIGNAL_SPOT_AZURE_SUBSCRIPTION_CONNECTED_SUCCESS=0
+export SIGNAL_SERVICE_PRINCIPAL_CREATION=0
 
 function init {
 	echo "{}" > $RESOURCE_FILE
@@ -23,7 +34,8 @@ function init {
 }
 
 function cleanup {
-	rm $POLICY_FILE_NAME_TEMP $POLICY_FILE_NAME $ACCOUNTS_STORE $SECRETS_STORE $TEMP_APP_DETAILS_LOCATION $ON_FAILURE_CLEANUP_FOLDER
+	log_debug "Cleaning up resources that have been used under /tmp dir..."
+	rm $POLICY_FILE_NAME_TEMP $POLICY_FILE_NAME $ACCOUNTS_STORE $SECRETS_STORE $TEMP_APP_DETAILS_LOCATION $RESOURCE_FILE_TEMP $RESOURCE_FILE
 }
 
 function format_timestamp {
@@ -127,8 +139,7 @@ function searchAndReplace () {
 	local filename=$3
 
 	echo "$1 $2 $3"
-	sed -i "" "s/{$search}/$replace/g" $POLICY_FILE_NAME
-
+	[ "$(uname)" = Linux ] && sed -i "s/{$search}/$replace/g" $POLICY_FILE_NAME || sed -i "" "s/{$search}/$replace/g" $POLICY_FILE_NAME
 }
 
 function validate {
@@ -168,12 +179,10 @@ function createAppRegistration {
 		registeredApplicationDetails==`az ad app create --display-name $chosenApplicationName --reply-urls "$SPOT_IDENTIFIER_URI" --available-to-other-tenants $APP_AVAILALBE_TO_OTHER_TENANTS`
 
 		case "$registeredApplicationDetails" in 
-	  		*"$APP_CREATION_SIGNAL"*) echo "Application Created Successfully"; echo $registeredApplicationDetails > $TEMP_APP_DETAILS_LOCATION; parseAzureFile $TEMP_APP_DETAILS_LOCATION; saveCreatedResourceToJsonObject "appRegistration" $(getAppId); break;;
+	  		*"$APP_CREATION_SIGNAL"*) echo "Application Created Successfully"; SIGNAL_APP_REGISTRATION_CREATION=1; echo $registeredApplicationDetails > $TEMP_APP_DETAILS_LOCATION; parseAzureFile $TEMP_APP_DETAILS_LOCATION; saveCreatedResourceToJsonObject "appRegistration" $(getAppId); break;;
 	  		* ) echo "Invalid, please run again.";;
 		esac
 	done
-
-	cat $TEMP_APP_DETAILS_LOCATION
 }
 
 function getAppId {
@@ -199,7 +208,7 @@ function validateCreatedRole {
 		spot_azure_created_role=`az role definition create --role-definition $POLICY_FILE_NAME --subscription "$account_to_connect"`
 
 		case "$spot_azure_created_role" in 
-	  		*"$APP_CREATION_SIGNAL"*) echo "Role Created Successfully"; break;;
+	  		*"$APP_CREATION_SIGNAL"*) echo "Role Created Successfully"; SIGNAL_IAM_ROLE_CREATION=1; break;;
 	  		* ) echo "Invalid, please run again.";;
 		esac
 		if [[ $attempts == 5 ]]; then
@@ -240,6 +249,7 @@ function connectSubscription {
 		echo $req
 		if [[ $statusCode == 200 ]]; then
 			log_info "Spot Azure subscription connected successfully" & sleep 1
+			SIGNAL_SPOT_AZURE_SUBSCRIPTION_CONNECTED_SUCCESS=1
 			jq <<< $req
 			break
 		fi
@@ -276,15 +286,10 @@ function assignRole {
 		role_name=$2
 		request=`az role assignment create --assignee $application_id --role $role_name`
 		case "$request" in 
-	  		*"$REQUEST_SUCCESS_CRITERIA"*) log_info "Successfully assigned IAM Role to Spot Azure Application"; CREATION_SUCCEEDED=1; saveCreatedResourceToJsonObject "roleAssignment" "$role_name"; break;;
+	  		*"$REQUEST_SUCCESS_CRITERIA"*) log_info "Successfully assigned IAM Role to Spot Azure Application"; SIGNAL_ROLE_ASSIGNMENT_SUCCESS=1; saveCreatedResourceToJsonObject "roleAssignment" "$role_name"; break;;
 	  		* ) log_error "Failed at $i attempt, retrying to assign role to application";;
 		esac
 	done
-
-	if [[ $CREATION_SUCCEEDED -eq 0 ]]; then
-		log_error "Couldn't assign role.. exiting the process"
-		exit 1
-	fi
 
 }
 
@@ -305,6 +310,10 @@ function execute_with_backoff {
 		if [[ $(echo $req | tr '[:upper:]' '[:lower:]') != *"error"* ]]; then
 			log_info "Successfully executed command. request output: $req"
 			IS_REQUEST_SUCCEEDED=1
+			if [[ "$SIGNAL_CRITERIA_SUCCESS_OUTPUT" != "$NO_SIGNAL_SUCCESS" ]]; then
+				echo "in cond"
+				 eval $SIGNAL_CRITERIA_SUCCESS_OUTPUT=1
+			fi
 			break
 		fi
 
@@ -321,15 +330,61 @@ function execute_with_backoff {
 }
 function special_execute() { set -x; "$@"; set +x; }
 
+function createIamRole {
+	local iam_role_name=$1
+	local account_to_connect=$2
+
+	curl -X GET https://spotinst-public.s3.amazonaws.com/assets/azure/custom_role_file.json > $POLICY_FILE_NAME
+
+	subscriptionId=`cat $ACCOUNTS_STORE | jq -r --arg account "$account_to_connect" '.[] | select(.name==$account).id' | tr -d '"'`
+
+	policyData=`cat $POLICY_FILE_NAME | jq '.properties'`
+	echo $policyData > $POLICY_FILE_NAME
+
+	# fix file according to Azure limits
+	searchAndReplace "subscriptionId" $subscriptionId $POLICY_FILE_NAME
+	searchAndReplace "customRoleName" $iam_role_name $POLICY_FILE_NAME
+	cat $POLICY_FILE_NAME | jq --arg roleName "$iam_role_name" '. + {Name: $roleName}' > $POLICY_FILE_NAME_TEMP
+	cp $POLICY_FILE_NAME_TEMP $POLICY_FILE_NAME
+	echo "" > $ACCOUNTS_STORE
+
+	log_info "Start creating IAM Role..."
+
+	execute_with_backoff "$SIGNAL_IAM_ROLE_CREATION_ENUM" "az" "role" "definition" "create" "--role-definition" "$POLICY_FILE_NAME"
+}
+
+function createServicePrincipals {
+	local application_id=$1
+	log_info "Assigning role to service principals" & sleep 10
+	execute_with_backoff "$SIGNAL_SERVICE_PRINCIPAL_CREATION_ENUM" "az" "ad" "sp" "create" "--id" "$application_id"
+}
+
+function validateCreatedResource {
+	if [[ ! -z $1 ]]; then
+		echo "yes"
+	fi
+}
+
+function resetSecrets {
+	local application_id=$1
+    newApplicationExpirationDate=$(getDate)
+    log_info "Resetting application to allow traffic from spot"
+	secret_details=`az ad app credential reset --id $application_id --end-date $newApplicationExpirationDate`
+	echo $secret_details > $SECRETS_STORE
+	if [[ ! -z $(cat $SECRETS_STORE)  ]]; then
+		SIGNAL_APPLICATION_SECRET_RESET=1
+	fi
+
+	# execute_with_backoff "$SIGNAL_APPLICATION_SECRET_RESET_ENUM" "az" "ad" "app" "credential" "reset" "--id" "$application_id" "--end-date" "$newApplicationExpirationDate > $SECRETS_STORE && sed -i '' '/^{/,/^}/!d' $SECRETS_STORE && sleep 2 && cat $SECRETS_STORE"
+}
 
 function handle {
-
 	azure_list_accounts=`az account list -o json`
 	log_info "These are the following available Azure accounts linked to your user, please choose one to connect to spot"
 	echo $azure_list_accounts | jq ".[].name" | tr -d '"'
 	echo -e "Which account (name) would you like to connect to Spot platform?"
 	read -r account_to_connect
-	echo -e "The command requires new registration creation in Azure Active Directory/App registration."
+	echo -e "The command requires new registration creation in Azure Active Directory/App registration.\n"
 
 	read -p "Continue (y/n)? or submit your application_id if already exists.. Choice: " choice
 
@@ -344,51 +399,38 @@ function handle {
 
 	log_info "appId : $application_id"
 
-    newApplicationExpirationDate=$(getDate)
+	if [[ $SIGNAL_APP_REGISTRATION_CREATION -eq 1 ]]; then
+		log_info "Application $application_id created successfully"
+		resetSecrets $application_id
+		if [[ $SIGNAL_APPLICATION_SECRET_RESET -eq 1 ]]; then
+			read -p "Please pick a name for the Spot for Azure IAM Role: " iam_role_name
+			echo $azure_list_accounts > $ACCOUNTS_STORE
+			createIamRole "$iam_role_name" "$account_to_connect"
+			echo $SIGNAL_IAM_ROLE_CREATION
+			if [[ $SIGNAL_IAM_ROLE_CREATION -eq 1 ]]; then
+				saveCreatedResourceToJsonObject "iamRole" "$iam_role_name" 
+				log_info "Start creating service_principals\n"
+				service_principal_details=`az ad sp create --id $application_id`
+				service_principal_dp=`echo $service_principal_details | jq -r '.displayName'`
+				if [[ ! -z $(validateCreatedResource $service_principal_dp) ]]; then
+					eval $SIGNAL_SERVICE_PRINCIPAL_CREATION_ENUM=1
+					if [[ $SIGNAL_SERVICE_PRINCIPAL_CREATION -eq 1 ]]; then
+						saveCreatedResourceToJsonObject "servicePrincipals" "$service_principal_dp"
+						assignRole $application_id $iam_role_name
+						if [[ $SIGNAL_ROLE_ASSIGNMENT_SUCCESS -eq 1 ]]; then
+							# store account details to connect spot account
+							echo $azure_list_accounts | jq -r --arg account "$account_to_connect" '.[] | select(.name==$account)' > $ACCOUNTS_STORE
+							connectAzureAccount
+							if [[ $SIGNAL_SPOT_AZURE_SUBSCRIPTION_CONNECTED_SUCCESS -eq 1 ]]; then
+								log_info "Successfully connected subscription to spot account."
+							fi
+						fi
+					fi
+				fi
+			fi
+		fi
+	fi
 
-    log_info "Resetting application to allow traffic from spot"
-	secret_details=`az ad app credential reset --id $application_id --end-date $newApplicationExpirationDate`
-	echo $secret_details > $SECRETS_STORE
-
-	# saveCreatedResourceToJsonObject "secrets" "$application_id" "$secret_details"
-
-	log_info "Start creating custom role...\n" && sleep 2
-
-	read -p "Please pick a name for the Spot for Azure IAM Role: " iam_role_name
-
-	curl -X GET https://spotinst-public.s3.amazonaws.com/assets/azure/custom_role_file.json > $POLICY_FILE_NAME
-
-	subscriptionId=`echo $azure_list_accounts | jq -r --arg account "$account_to_connect" '.[] | select(.name==$account).id' | tr -d '"'`
-
-	policyData=`cat $POLICY_FILE_NAME | jq '.properties'`
-	echo $policyData > $POLICY_FILE_NAME
-
-	# fix file according to Azure limits
-	searchAndReplace "subscriptionId" $subscriptionId $POLICY_FILE_NAME
-	searchAndReplace "customRoleName" $iam_role_name $POLICY_FILE_NAME
-	cat $POLICY_FILE_NAME | jq --arg roleName "$iam_role_name" '. + {Name: $roleName}' > $POLICY_FILE_NAME_TEMP
-	cp $POLICY_FILE_NAME_TEMP $POLICY_FILE_NAME
-
-	log_info "Start creating IAM Role..."
-
-	spot_azure_created_role=`az role definition create --role-definition $POLICY_FILE_NAME`
-
-	saveCreatedResourceToJsonObject "iamRole" "$iam_role_name" 
-
-	log_info "Assigning role to service principals" & sleep 10
-		
-	# create SP for app
-	service_principal_details=`az ad sp create --id $application_id`
-	service_principal_dp=`echo $service_principal_details | jq -r '.displayName'`
-
-	saveCreatedResourceToJsonObject "servicePrincipals" "$service_principal_dp"
-
-	assignRole $application_id $iam_role_name
-
-	echo $azure_list_accounts | jq -r --arg account "$account_to_connect" '.[] | select(.name==$account)' > $ACCOUNTS_STORE
-
-	connectAzureAccount
-	log_info "Finished"
 
 }
